Updated to support multiple /home directories
This version is exactly the same as Boom Shadow’s version, however, I’ve modified it to support additional /home directories so that permissions extend beyond just the default /home directory.

In our servers, we often have a surplus of RAM and CPU due to governors that we’ve put in place to protect and “keep fair” the resources on our servers, but we eventually eat up disk space before we reach a point of maxing out RAM or CPU. So we use cPanel’s ability to create additional home directories (ie., /home, /home2, /home3), each of them on their own attached storage volume.

You’ll run the script the same way, but this version of the script will automatically find the user’s “true” home directory. Once it finds that, it will continue as this script has always operated…fixing the permissions.

- finds the user's true home directory
- runs all the same great commands on that directory
- run the script as you normally would, nothing changes

Signed-off-by: Will Ashworth <will@ashworthconsulting.com>