Update fixperms.sh
This fixes and edge case where hidden files or folders are ignored. Seen a few accounts i ran this on cPanel wise not fixed entirely and fixed it in my personal cPanel/CyberPanel copy. I'm submitting this upstream so it saves some people a real headache later on as a courtesy.

Reference:
https://github.com/whattheserver/cyberpanel-fixperms/issues/1